DB-4822: Moving over BulkWrite change from DB-4693.
In K2, we had already solved hanging import issues that were due to
regions splitting and caches etc. On a lark, I pulled those changes over
manually to Lassen (modifying to fit the different architecture), and
tried them out. I was able to load the lineitem table from TPCH1000 on a
cluster with no problems once that change was made.