DB-4864: Presumed abort.
When a transaction record is missing, it's pretty much always a sign of
database corruption in one form or another. In some particular cases in
the DataDictionary, it was possible for some Puts to be written without
the underlying Transaction Id as its timestamp. When that happens, HBase
uses the current system time on the server instead. This will invariably
result in a failure when reading or writing to that row because the
transaction with the current timestamp does not exist.

This commit finds all the places where this error was present (I hope),
but once written, the damage is done, and the only way to avoid the
problem is to reload the data.

To avoid requiring this of large data environments (i.e. HarteHanks, McKesson),
this commit also implements a "presumed abort" strategy. Whenever a transaction
record cannot be found by the server, we will emit an error message to the log,
but treat the transaction as rolled back instead of failing. This means that queries
will not fail due to corrupted transaction data, but has the unintended
side effect where a transaction being created with a previously missing
data point will cause bad data to suddenly appear. This is clearly very
bad, but also very unlikely to happen, since current timestamps are in
the trillions (or more), whereas databases will have not reached those
transactions for a very long time. At least, I hope