DB-4822: Correcting race condition with multiple region moves.
If you consider a sequence of events using two separate region servers
and a single region as follows:

1. move region from Server A to Server B
2. move region from Server B to Server A

If this happens under sufficient load, then it is possible to encounter
a race condition where the write pipeline continues to refer to a region
which is "closed" (because of the first move) even though that region
was garbage collected. This amounts to a resource leak on the old region
object, but it manifests as a permanent cycle where rows are sent back
with a NotServingRegion error code forever.

The core problem was because of an attempted "clever" optimization,
where we attempt to use reference counting to ensure that a given object
is reused. That object is not particularly expensive, however, so there
is no reason not to create a new one each time it's requested
(especially since it's being held on to). This fix removes the reference
counting and just creates a new instance each time; when a region is
moved or closed, then the reference is garbage collected and the problem
dissappears.