DB-4184: Case-sensitive table and schema names.
It is possible to create tables and schemas with mixed case; this means
we need to be able to collect statistics on those. It's mainly a matter
of obeying case-sensitivity when parsing table and schema names during
stored procedures.

I modified the "EngineUtils.validateSchema" and
EngineUtils.validateTable" commands to correctly deal with mixed-case
tables, and made sure that they were being used everywhere. I also added
two ITs to verify.