DB-4684: Ignoring TableNotFoundExceptions when vacuuming an old conglomerate.
If you have two concurrent Vacuum's going on, then one may delete a
table before the other can get to it. When that happens, you'll get a
TableNotFoundException from the HBaseAdmin. In that case, there isn't
much to worry about, since a vacuum is intending to delete a table
anyway. We just go ahead and ignore the error.