DB-3626: Correcting TimeZone issues in test.
It turns out that we were using JodaTime to construct new date objects.
This is okay in some cases, but is apparently not good with timezones.
The whole reason it was done was to avoid using a Calendar object during
imports; that was mainly because we were unable to properly cache the
expensive to create Calendar object. I re-wired the AbstractFileFunction
to keep a cached Calendar object; with a cached object, the need to use
JodaTime is eliminated and so we can go back to the old derby approach,
which is correct for all time zones.