DB-5119: Adding in TxnFilter reset in constraint checking.
Turns out we were missing a reset of the transaction filter, so a delete
of a prior row was contaminating attempts to insert into that same row.
This adds the reset in the filter. It unfortunately doesn't make much
sense to add this to the transaction UTs, because there isn't really a
concept of updates in those UTs, so it wouldn't really reveal the
problem. An IT which performs updates will have to do instead.