DB-4929: Correcting shallow-clone error.
We were making shallow clones of Date,Time,and Timestamp columns when
constructing statistics. This meant that we were effectively sharing a
mutable, non-thread-safe object (the HyperLogLog cardinality element)
among multiple threads. This can cause all kinds of nasty errors.

This also fixes an accidental mutable array share during clone of
SparseHyperLogLog instances that could have resulted in other errors.
These are highly unlikely because we generally only use merge() and
clone() methods after we've finished performing writes, so we are
unlikely to use the shared arrays in an unsafe way, but this is still a
safer approach to cloning.