DB-5367: Able to create a child Transaction of a SpliceTransactionView.
This makes it possible to create a child of a SpliceTransactionView,
as long as the writability semantics are maintained from parent to
child.  That is, you can only create a writable child of a
SpliceTransactionView if the parent is also writable
(otherwise, you lack the information to record data in the TXN table).

This is a manual re-application of git commit
be2ea02ea5e1af3ee39610bfe7ee34f83ac019f0
in master-k2.