DB-4324: Correcting transactional typo with deletes and updates.
It turns out, if you delete a row, rollback the delete, then perform an
update, the update corrupts the table, because the insert columns go
missing. This is just because SITransactor was using
"Txn.conflicts(deleteTxn)" to determine whether or not to add an
AntiTombstone during writing. Since the delete was rolled back, there is
no conflict, so the transactor wanted to add an anti-tombstone to the
cell. When we go to read the row then, we stop at the anti-tombstone
because we "know" that everything below it was deleted. Thus we lose the
insert.

The correct way to determine whether or not an anti-tombstone should be
placed is by using the "Txn.canSee(deleteTxn)" method. This determines
whether or not the transaction is _visible_ to the update, not just
whether or not it conflicts.

This is actually unit-testable, so this commit includes an IT (in
ForeignKey Code) and a unit test in TransactionInteractionTest.