DB-5058: Correcting copy error in RowOrderingImpl.
We weren't copying unordered optimizable information in RowOrderingImpl
when using the copyTo() method. This meant that unordered optimizable
information was lost as we moved from conglomerate to conglomerate (and
through join strategies). As a result, we ended up believing that we
didn't need a sort even when we did.

This patch copies over the unordered optimizable information and adds an
IT to check on it.