DB-5228: Carrying over 1.5.x index fix for updates.
When issuing an update to an indexed column, we issue a delete and then
an insert. If we are updating in such a way as to insert into the same
row (i.e. updating a value with itself), then we end up with an insert
and a tombstone at the same cell value. When that happens, the tombstone
wins and the data goes missing from scans.

This was discovered and resolved in 1.5.x (git revision
9fba1f601b8a8f0db17bf5d2634885f417029a71), so this just carries over the
fix from 1.5.x and adds a new IT to verify.