DB-5515: Protecting against weird HDFS permissions error preventing import.
The error can't be reproduced locally--most likely we need a cluster (and the
DistributedFileSystem) to make this work. However, just examining the
stack trace: the call being made is not essential to the process of
opening the bad directory. The usage of the method itself is in a
location where we don't necessarily NEED to the results of that
operation. To that end, we simply catch and swallow the error (logging a
warning).

Unfortunately, since it can't be reproduced locally, we can't write an
IT to verify it.