SPLICE-868: Correcting Race condition around splits.
It turns out that a Region is not to be considered "available" until it
is added to the HBase RegionServerServices, instead of when a
coprocessor start() is called. For most operations this doesn't matter,
because RegionServerServices is where the HBase network layer determines
if the Region is present or not; if you sent an operation to that region
and it wasn't in RegionServerServices, you'd get a
NotServingRegionException instead of allowing through an operation.

However, in the Write Pipeline we send writes to multiple regions
simultaneously in a single network call (if there's a network call at
all--in some cases we send the data directly to the same JVM where we
are writing data). Therefore, we allow the possibility of bypassing the
network stack and its protections.

When operating outside the stack, there is a race condition around
splits where a daughter region starts all of its coprocessors up before
it is finished fully initializing (in fact, start() is called on all
coprocessors BEFORE preOpen() is called on any RegionObserver). This
means that we were allowing some writes to proceed to a daughter region
before it was fully opening, leading to a race condition and (in turn)
causing the abort sequence to occur.

This corrects that issue by requiring that the region be present in the
RegionServerServices before allowing a BulkWrite to proceed.

The awkward raciness of this prevents a decent IT; thankfully, Jenkins
will fail periodically with an ABORT message during builds if this
regresses, which will serve as a decent IT (since it breaks the build).