DB 5753 : backport 2.5
Added check constraint during the table definition.

Now check the number of attributes to match what we have in the external file
Now check the datatypes to match the external file
clean up tests

Add Error Handling for ORC and Parquet Files where the definition does not match the data in file.

WIll create a 2.5 commit after this one is in master