SPLICE-1370 ; INSERT, UPDATE, DELETE error message for pin tables
Fixed the issues :
INSERT INTO the pinned view reports success and inserts into the original table (but not the pinned view). This should not be allowed.
DELETE FROM the pinned view reports '0 rows inserted/updated/deleted', but does not display an error message, which is misleading
UPDATE on a pinned view generates a network protocol error and bombs Splice