Overhauled markdown
Brought in a fresh copy of the docs markdown and converted it to work with GFM. The html parser in GFM struggles with some constructs, which throws off indentation, e.g. makes numbered lists wrong by indenting and changing the style of items. I've worked to fix all of these.