SPLICE-2228 Fix start-splice-cluster ZooKeeper liveness check
Changed `until` condition to stop when `COUNT` equals `0` so that the
condition following the loop will evaluate to true, as expected, when
ZooKeeper is not found to be alive.

Also, changed condition to break out of the loop to succeed when `nc`
returns a non-empty string instead of an empty string.