SPLICE-2228 Fix ZooKeeper liveness check in scripts (#2175)
* SPLICE-2226 Fix standalone ZooKeeper liveness check

Changed `until` condition to stop when `COUNT` equals `0` so that the
condition following the loop will evaluate to true, as expected, when
ZooKeeper is not found to be alive.

Also, changed condition to break out of the loop to succeed when `nc`
returns a non-empty string instead of an empty string.

* SPLICE-2228 Fix start-splice-cluster ZooKeeper liveness check

Changed `until` condition to stop when `COUNT` equals `0` so that the
condition following the loop will evaluate to true, as expected, when
ZooKeeper is not found to be alive.

Also, changed condition to break out of the loop to succeed when `nc`
returns a non-empty string instead of an empty string.

* SPLICE-2228 Fix ZooKeeper liveness check in shell scripts

Simplified and corrected ZooKeeper liveness check in all shell scripts
that include the check. In addition, corrected ZooKeeper log file
reference in the error message that is issued when it is determined that
ZooKeeper is not live.

* SPLICE-2228 Fix ZooKeeper liveness check in shell scripts

Check ZooKeeper response to `ruok` is exaclty `imok` rather than any
non-empty string. This ensures this will fail when we upgrade to future
versions of ZooKeeper where 4lw whitelisting is added, forcing
corresponding changes to these scripts.