DBAAS build 1845
Added `platform_it` module as a dependency to the `assembly` module (for
hregion servers) and the `splice_spark` module (for spark executors) due
to dependency refactoring.  Also, added back the relocation of netty
because another module expected it to be relocated.

Updated imports for some parquet-column classes due to migration from
Twitter's parquet-column artifact to Apache's, which caused the base
package to change from `parquet` to `org.apache.parquet`.

DBAAS-1897