DB-8793 Remove deleted rows during compaction
We can remove deleted rows only if active transactions are using a more
recent version of that row.

In the event of a minor compaction, since we do not see all hfiles and
might have a partial view of the history of a given row, we delete
everything BUT the tombstone itself. This way, the portion of history
that lived in a different hfile will still fall earlier than the
remaining tombstone, ensuring that the known state of the row is
"deleted".
Those remaining tombstones shall be removed during major compaction.