DB-9448 Fix resurfacing rows issue in Spark
Migration to hbase2.0 created two issues:
1. We did not use memStoreAware's lock properly anymore
2. We did not know about compacted files when creating a HRegion from
Spark's side

This commit addresses these two issues in the following way:

1.
SpliceCompactionRequest.setOffPeak is now hijacked to properly
release the lock on memStoreAware once a compaction has completed.
Short of modifying HBase's code to be able to use a custom
CompactionLifeCycleTracker (only DUMMY is used outside of their tests),
this seems to be the only reliable solution.

2.
Spark side now RPC calls the relevant region to get a list of compacted
files so that regions created on Spark side can properly exclude
compacted store files that pertains to a specific region.